+++ src/contract/internal/index.js
@@ -3,6 +3,7 @@ const SponsorWhitelistControl = require('./SponsorWhitelistControl.json');
 const Staking = require('./Staking.json');
 const ConfluxContext = require('./ConfluxContext.json');
 const PoSRegister = require('./PoSRegister.json');
+const CrossSpaceCall = require('./CrossSpaceCall.json');
 
 module.exports = {
   AdminControl,
@@ -10,4 +11,5 @@ module.exports = {
   Staking,
   ConfluxContext,
   PoSRegister,
+  CrossSpaceCall,
 };